Fix build when `BUILD_MODEL_REDUCTION` is ON.
Rename `BUILD_MODEL_REDUCTION` to `VEGA_BUILD_MODEL_REDUCTION`.
Add some include directories for "private" headers to get picked
up by dependent libraries; for example, the quaternion library
now marks its own source directory (`src/libquaternion`) as an
include directory for targets that depend on it. Note that this
is only for the build tree, not the installed version of Vega.

There is still some fiddliness going on with LAPACK; at times
CMake does not seem to want to link to it even though it is listed
as a dependency. The workaround seems to be upgrading to CMake 3.1.